+++ ChangeLog.md
@@ -2,6 +2,8 @@
   * Core
     * Add ship role 'Taxi'
     * Rename 'Companion App' tab to 'Frontier API' and update relevant text to clarify its use and operation
+  * Events
+    * Add 'Settlement approached' event
 
 #2.2.0
   * Core